V0.2.0
Criado sistema de salvamento de Layouts no Banco de Dados

Criado sistema de carregamento de Layouts do Banco de Dados.

Desmembrado o arquivo database_helper: O Banco de Dados dos Layouts foi para o database_layout, e o Banco de Dados dos Usuários foi para o database_users.

Criado menu para seleção de criação de novo layout ou carregamento de layout já existente.

Criado widget de nome do Layout e ajustado para aparecer o nome do layout carregado no topo da tela.

+++ lib/ui/layout_dos_cartoes/slide.dart
@@ -22,7 +22,7 @@ class _SlideCircleState extends State<SlideCircle> {
   @override
   void initState() {
     super.initState();
-    circleSize = widget.circleSize; // Valor inicial do slider
+    circleSize= widget.circleSize.toDouble(); // Valor inicial do slider
   }
